+++ tests/Plugin/Wechat/Virtual/AddPayloadSignaturePluginTest.php
@@ -11,11 +11,6 @@ use Yansongda\Pay\Plugin\Wechat\Virtual\AddPayloadSignaturePlugin;
 use Yansongda\Pay\Tests\TestCase;
 use Yansongda\Supports\Collection;
 
-/**
- * @internal
- *
- * @coversNothing
- */
 class AddPayloadSignaturePluginTest extends TestCase
 {
     protected AddPayloadSignaturePlugin $plugin;