+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## Unreleased
 
+
+## v3.8.0-beta.2 - 2026-07-01
+
 ### Added
 
 - 新增微信小程序虚拟支付支持 (#1172)